fix: handle paginated API responses and fix session expiry handling
- quotasApi: extract .items from paginated response in getQuotasByTransaction
  (DebtForecastScreen was crashing on spread of {items, metadata} object)
- quotasApi: move Quota/QuotaWithTransaction to types/quota.ts and re-export
- quotasApi: remove duplicate PaginatedResponse (canonical in creditCardsApi)
- useAuth: add SessionExpiredError class; requestWithAuth throws instead of
  returning a fake 401 so React Query retry logic can detect session expiry
- _layout.tsx: configure QueryClient to not retry on SessionExpiredError
- Fix import group ordering in modified files (external before internal)
