+++ package.json
@@ -45,6 +45,7 @@
     "react-dom": "19.1.0",
     "react-native": "0.81.5",
     "react-native-chart-kit": "^6.12.0",
+    "react-native-device-info": "^14.1.1",
     "react-native-gesture-handler": "~2.28.0",
     "react-native-modal": "^14.0.0-rc.1",
     "react-native-reanimated": "~4.1.1",
